Rename method.
